+++ lib/kitchen/directions/chapter_review_container/main.rb
@@ -2,7 +2,7 @@
 
 module Kitchen
   module Directions
-    module BakeChapterReview
+    module ChapterReviewContainer
       def self.v1(chapter:, metadata_source:)
         V1.new.bake(chapter: chapter, metadata_source: metadata_source)
       end