+++ spec/directions/chapter_review_container/main_spec.rb
@@ -2,9 +2,9 @@
 
 require 'spec_helper'
 
-RSpec.describe Kitchen::Directions::BakeChapterReview do
+RSpec.describe Kitchen::Directions::ChapterReviewContainer do
   it 'calls v1' do
-    expect_any_instance_of(Kitchen::Directions::BakeChapterReview::V1).to receive(:bake)
+    expect_any_instance_of(Kitchen::Directions::ChapterReviewContainer::V1).to receive(:bake)
       .with(chapter: 'chapter1', metadata_source: 'metadata')
     described_class.v1(chapter: 'chapter1', metadata_source: 'metadata')
   end