+++ spec/directions/move_solutions_to_answer_key/main_spec.rb
@@ -2,9 +2,9 @@
 
 require 'spec_helper'
 
-RSpec.describe Kitchen::Directions::BakeChapterAnswerKey do
+RSpec.describe Kitchen::Directions::MoveSolutionsToAnswerKey do
   it 'calls v1' do
-    expect_any_instance_of(Kitchen::Directions::BakeChapterAnswerKey::V1).to receive(:bake)
+    expect_any_instance_of(Kitchen::Directions::MoveSolutionsToAnswerKey::V1).to receive(:bake)
       .with(chapter: 'chapter1', metadata_source: 'metadata', strategy: 'strategy', append_to: 'element')
     described_class.v1(chapter: 'chapter1', metadata_source: 'metadata', strategy: 'strategy', append_to: 'element')
   end